[메인] fix : swiper js 분리 및 페이지네이션 버그수정
